Stop tracking the backup directory
We will instead using the stage/ directory to keep new packages. All built
packages will be saved to this directory so that the packager can easily
inspect the output before adding the packages to the repository. Once it is
determined that the packages are correct, package.py can be used to add
selected packages to the repository.

This directory will be sync'd with demizerone.com so users can inspect the
packages contained in the stage if needed.

The backup directory was extra work for no reason as the packages could easily
be regenerated from the git history if needed.