Find dataset when multiple pools exist on boot
If the `zfs` kernel parameter is set to `bootfs` and multiple pools
are available, check each pool for the bootfs property.

As far as I am aware, there are only three possible strings that might
be returned by `zpool list -H -o bootfs`, an empty string, `-`, and
a dataset name. If this is not the case this change has the potential
to break existing systems.