repo.sh: Fixed missing package list bug
If the either the i686 or x86_64 packages are missing, an attempt was still
made to copy the packages using the cp_list string to the non existent
destination repo. This caused the script to terminate on error of cp with no
destination repository. The fix was to check cp_list for "". If cp_list is
empty, there are no packages to copy and the for loop is continued to the next
architecture.