scraper.sh: Fix curl pipe problem
Thanks to the switch from explicitly sending mail to relying on cronie,
I (hopefully) discovered the cause of a long occurring problem with curl
that was causing false positives with version checking. When curl
output was piped to grep, grep (for some reason) was not reading the
full output, causing curl to return an error "(23) failed writing body
pipe". This was fixed by saving curl output to a variable an piping that
to grep.