initscript refactor and test
Init scripts have been refactored to improve readability. The most
important aspect of this refactor is that automated testing has been
implemented using Qemu. Eventually we will get to the point of
automatically building new archzfs packages as soon as a new kernel is
released to the arch testing repo, after that the tests will be used to
determine if the archzfs packages can be pushed out to aur and
archzfs.com.

Closes #22
Closes #31
Closes #38
Closes #48