Add support for archiso packages
Packages built just for the archiso, included in the archzfs main repo.

To install, users will only need to install the archzfs-archiso-linux
package group.

NOTE: This will not work with the default clean-chroot-manager as the
required kernel packages for archiso cannot be installed in the clean
chroot working copy because they are out-of-date. The following change
was made to ccm64 to build the archiso packages:

    PACKAGER=$PACKAGER THREADS=$THREADS nice -19 /usr/bin/makechrootpkg -I "/var/cache/pacman/pkg/linux-4.5.4-1-x86_64.pkg.tar.xz" -I "/var/cache/pacman/pkg/linux-headers-4.5.4-1-x86_64.pkg.tar.xz" -c -u -r $CHROOTPATH64

A request has been made upstream to pass -I to makechrootpkg so that
this change no longer requires a custom ccm64 script.