Experimental support for extramodules
Allows for building the archzfs packages once per major kernel version.

This will make it easier to support these packages and less pain when
updating for users of these packages.

Also included in this change is the ability to install archzfs-linux and
archzfs-linux-lts (not archzfs-linux-git) at the same time. This is done
by adding new header packages: spl-linux-headers and zfs-linux-headers.

NOTE: upstream ZFSOnLinux developers recommend re-compiling on every
kernel version change. This is the norm for kernel modules. In my five
years of maintaing this package, the builds have never broken on minor
kernel updates (4.9.7 -> 4.9.8), therefore, this change is marked as
experimental until it has been determined nothing breaks.

Closes #92
Closes #87
Closes #75