Mount ZFS_DATASET regardless of canmount=noauto
The root dataset should be mounted even if canmount is set to noauto.
This is because there can be multiple root datasets if one is using
boot environments using a tool like zectl (https://github.com/johnramsden/zectl).
The root datasets always have the same mountpoint (/), so you cannot
set canmount=on on all of them as that would result in overlapping mounts.