Support overriding encryption keylocation to prompt in initcpio hook
If an encrypted ZFS dataset specifies a file:// keylocation, relying on
the default `zfs load-key` behavior will prevent the user from typing a
passphrase even if its keyformat is passphrase. This renders an initcpio
that does not contain the key file unbootable unless SSH has been
preconfigured and the user connects remotely do unlock the dataset.

To allow some recovery, the hook now will attempt to force a prompt if
the default load-key behavior failed to unlock the dataset and the
keyformat is passphrase. This is still compatible with providing a
key passphrase on, e.g., some USB device because the user can insert the
device after the fact and then just interrupt the load-key prompt (or
enter a garbage password three times) to retry the default load.

Also, zfs_decrypt_fs has been reworked to monitor the return value of
ZFS operations, which allows graceful failure for ZFS versions that do
not support encryption without having to scrape the help text of zfs.