Workaround for 'common.sh: No such file' error
File /usr/share/zfs/common.sh is installed by zfs-utils, however zfs install build target will also try to build it with other packages.
The failure to create this file does not break the build, but it gets in a way of diagnosing other build problems, so best get rid of it.
Here we create the directory to store this file (so sed instruction can create it) and, after the make install is completed, we remove it
 - so it does not get in the way of the same file installed by zfs-utils.