Disable exit on error while preforming builds.
A failure in any of the kernel builds will cause the build script to
exit. Stable outruns OpenZFS regularly and it often takes time for
OpenZFS to allow builds against new versions of the kernel. This means
that lts will not be rebuilt for long periods of time since the build
script breaks whenever a major version bump happens to stable.