ci: more resilient releasing, fixes #592
Our releases use fixed names like "experimental", which we constantly
update. Because existing releases are updated "in place", there's a
slight risk that we get a broken Pacman repo, e.g. #607.

This commit introduces the idea of #592: releases are created first
under a temporary name before being moved to the final name. Thus, if a
new release breaks, it should not affect the existing release.

This solution is probably not entirely bulletproof either, but will
hopefully be an improvement over the current situation.